ch4: add a generic hint for vci selection
This hint controls both sender and receiver VCI. Its use can be
extended for collectives, e.g., to select a progress thread
to assign/offload the operation to.

Details:
Specific sender and receiver VCI can be selected for each communicator
using this hint. Applications can utilize this feature to control how
messages are spread across the VCIs. For instance, programmers can write
the applications so that pairs of threads communicate by using different
communicator. Then each communicator can be mapped to a different VCI
through the thread_id info hint. Valid values for this info hints are 0
to (total vcis - 1).