ch4/ofi: add barrier during vni address exchange
Ranks should wait until address exchange is complete and the av
table is setup in all the ranks.

Otherwise, some processes may try to access the av table prematurely,
e.g., when different ranks exit address exchange at different times.
Such problem was observed with the sockets provider. One rank was still
under address exchange, when it received an incoming communication
from a rank that already completed its address exchange steps. Then
libfabric processed this communication for the first rank, and tried
to to an av lookup for the remote rank, but before that  address
added to the av. This resulted in test failures.