ch4: add CVAR for auto affinity of async progress threads
This allows user to toggle the auto affinity of the async progress
threads. If this CVAR is true, MPICH will automatically select an
affinity of the async progress threads to the CPU cores. If false,
affinity is determined by other mechanisms when present, e.g.,
MPIR_CVAR_CH4_PROGRESS_THREAD_AFFINITY. In the absence of such
additional mechanism, cpu core affinity is not applied to the async
progress threads.